xcm: move pallet xcm benchmark to xcm_config.rs
